bug: remove mislabeled 1.56 firmware:
"It does look like both of those files are the same. I looked through my archive and I do not have a different version for 1.56 so it seems like that version may be lost to into the void. Version 1.57 should be the correct version to use." - Eric Maycock (CTO of Inovelli)